+++ js/main.js
@@ -1,7 +1,9 @@
 import Phaser from "phaser";
 import { pinstage1 } from "./pin-stage1";
+import { pinstage2 } from "./pin-stage2";
 import { towerstage1 } from "./tower-stage1";
 import { startMenu } from "./start-menu";
+import { pinstage2 } from "./pin-stage2";
 
 const D_WIDTH = 1000;
 const D_HEIGHT = 600;
@@ -14,7 +16,7 @@ export const config = {
   width: D_WIDTH, // ゲーム画面の横幅
   height: D_HEIGHT, // ゲーム画面の高さ
   antialias: false, // アンチエイリアスを無効にするらしい（分ってない）
-  scene: [towerstage1,startMenu, pinstage1, ],
+  scene: [towerstage1,startMenu, pinstage1, pinstage2],
 
 
   fps: {
