+++ js/pin-stage1.js
@@ -170,6 +170,8 @@ this.scene.restart();
 
   var graphics = this.add.graphics(); //暗転用のグラフィックス
   var gameoverText;
+  var gameclearText;
+  var nextText;
   var restartText;
   var returnMenuText;
   // 画面全体に配置
@@ -198,7 +200,18 @@ this.scene.restart();
   }
   //人間と宝がぶつかったときの処理
   function hittreasure(humanImage, treasure) {
-    var gameclearText = this.add.text(220, 70, "GAME CLEAR", redtext); //ゲームクリアの表示
+    gameclearText = this.add.text(220, 70, "GAME CLEAR", redtext); //ゲームクリアの表示
+    gameclearText.setDepth(1);
+    nextText = this.add.text(390,200,"次へ",whiteText);
+    returnMenuText = this.add.text(420,300,"ホーム",whiteText);
+    nextText.setInteractive();
+    returnMenuText.setInteractive();
+    nextText.on("pointerdown",() => {
+      this.scene.start("pinstage2");//次のステージへ移動する処理
+    });
+    nextText.setDepth(1);
+    graphics.setDepth(1); // 暗転用のグラフィックスを前面に表示
+    returnMenuText.setDepth(1);
   }
 
   let pinsClicked = 0; //クリックされた画像の数（pin1とpin2のみ）
